requirements.txt
streamlit>=1.30.0
openai>=1.0.0
PyMuPDF>=1.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
